Add support for generating completion scripts for shells other than bash
This change adds support for generating completion scripts for shells
other than bash to the shell-complete utility program. Specifically, it
now expects the shell to generate such a script for as its first
positional argument.
We do not add tests for the completion functionality for the newly
supported shells. Such tests are a pain to write, they are unlikely to
cover anything that the clap test suite does not already, and we simply
do not have all such shells available.