Allow spaces in base32 OTP secrets
Some sites, such as GitLab, display their OTP secrets as base32 strings
in groups separated by spaces.  Previously, the user had to manually
remove the spaces before passing the secret to nitrocli.  But as the
space can never be part of the base32 string itself, we can
unambiguously identify it as a separator and remove it from the secret
before processing it.

Fixes #136