Add pws update subcommand
This patch adds the pws update subcommand that overwrites some fields of
a programmed PWS slot. We always have to specify the values for all
fields when calling writing to a slot, so we first query the current
values of the fields that remain unchanged and then execute the write
command. This is a potential race condition, but concurrent access to a
Nitrokey device is unsupported and highly likely to mess up the
communication anyway.