Change error formatting to omit backtraces
The format specifier we use for printing errors returned by the main
program will include a backtrace in the output -- if one is available.
That's not necessarily a feature that we need: the causal chain of
errors should be sufficient and end users are unlikely to have use for a
backtrace.
This change adjusts the format specifier so that we only print the chain
of errors, albeit in a slightly different format:
- Previously:
  > Failed to generate OTP
  >
  > Caused by:
  >     Command error: The given slot is not programmed

- New:
  > Failed to generate OTP: Command error: The given slot is not programmed