Allow serial number filtering for Nitrokey Storage
The Nitrokey Storage does not report a serial number in the USB device
descriptor. Because of that we have not supported filtering of those
devices by serial number. However, by actually connecting to the device
we can still retrieve such a number and, in fact, that's what we report
as part of the list command.
With this change we mirror this logic in the device discovery/filtering
pass: if the reported device information does not contain a serial
number (as is the case for Nitrokey Storage), we connect to the device
and look it up this way. In so doing we reduce the amount of confusion
surrounding the --serial-number option and configuration file argument.