Support reading PWS passwords from stdin
Specifying passwords as command-line arguments can be problematic:  The
user has to be careful that the command is not recorded in the shell
history.  Also, the password might be exposed over the process list.

This patch adds an alternative input method for the pws add and pws
update methods: reading the password from stdin.  This is activated by
setting the password to "-".  For the tests, we use a string as a
simulated stdin buffer.