Make otp-cache a binary target of nitrocli
Upon its inception, the common nitrocli extension glue code was put into
a crate, nitrocli-ext. Doing so comes with a slew of challenges in terms
of distribution, versioning, and compatibility with the main program.
Because a different crate is used, that property transfers to
extensions, meaning that they also have to be separate crates, for
example. Packaging of Rust programs on some Linux distributions is based
on content downloaded from crates.io, effectively forcing us to publish
the extension crates, which in turn forces us to publish nitrocli-ext
and to commit to a stable API. Then we get into questions of how to
version separately published creates and how to ensure or enforce
compatibility. In short, nitrocli-ext being a crate has trickle down
effects that introduce additional complications that we did not
anticipate initially.
To that end, with this change we make nitrocli-ext a mere module of the
crate and the otp-cache core extension a binary target. This implies
that both are versioned as the main program is and are also available
when pulling the published archive from crates.io, circumventing most if
not all of the aforementioned problems.