Remove CI badge from Cargo.toml
The CI badge that we specified in Cargo.toml is no longer used by
crates.io and being phased out. It is also no longer valid since we
switched from using GitLab as the CI provider to GitHub Actions. As
such, there is no point in keeping the entry and this change removes it.