Adjust extension argument test to expect full variable names
In the future we want to test the existence of some environment
variables that are not prefixed with 'NITROCLI_'. The test logic that we
currently have in place implicitly assumes such a prefix, however. To
fix this problem, this change reworks the test extension script to stop
any implicit prefixing and instead expect the full environment variable
name to be provided to it. As a side effect, that makes the logic
simpler.