Factor out TTY logic into separate module
The logic we have in place for retrieving the path to the TTY to set
the GPG_TTY environment variable to if it is currently unset is rather
complex and very specific to the operating system used. In addition, for
better or worse we have plans to extend further, necessitating the usage
of even more OS specific functionality.
In order to not pollute the pinentry module with too much of this logic,
this change moves this functionality into a new module, tty.
Furthermore, in an attempt to maintain a clear separation between
the support code for different operating systems, we split this module
into different sub-modules that are compiled conditionally.