Adjust retrieve_tty to return anyhow errors
When we devised the retrieve_tty function we skimped on the error
reporting because 1) we never actually evaluate the error in any way and
2) it's just a major pain to work with errno (in general, but in Rust
specifically).
However, now that we have simplified the function down to a single call
into the standard library we may as well do it right and bubble up
errors appropriately, because, even if not used most of the time, it can
still help tremendously with debugging. This change does so by
manifesting an anyhow error, as we do in most parts of the crate.