Walk process tree as part of TTY inference
Unless the GPG_TTY environment variable is set by the user, we try our
best to find the process' TTY ourselves and set the environment variable
accordingly. However, so far we only looked at our own process' stdin to
see whether it refers to a TTY.
Alas, we could have been invoked by another process and stdin could have
been redirected arbitrarily in such a case. git-send-email, for example,
does such a dance. If that's the case, we'd end up reporting a file as a
TTY candidate that isn't really a TTY.
It seems somewhat sensible to just "do the right thing" and walk up the
process hierarchy until we find "the" TTY to use (at least until we get
a report where that breaks). With this change we do exactly that.