Add Nix flake
This commit embeds a Nix flake into the repository allowing a Nix user
to run nitrocli by issuing:

    nix run d-e-s-o/nitrocli?dir=contrib/nix

This commit also adds a reproducible way to install nitrocli on a system
managed by Nix flakes and to start a development environment.

The installed package includes man pages and shell completions.