Statically linked binary
Cross compiling directly on the target system for musl seems to hit some
snags with respect to the C++ code we use. The problem may be with
libnitrkey's build system or perhaps nitrokey-sys-rs. Unclear.

One possible avenue may be to try and link libstdc++ statically. gcc
seems to support -static-libstdc++ for that purposes.

Ideally that would be the avenue we pursue and get working.

Using cross-rs (which does everything in a musl based container) seems
to circumvent that and gets us further. However, there we have the
problem that we do not have hidapi-usb available and it's unclear how to
get it there.
The ubuntu package contains
/usr/lib/x86_64-linux-gnu/libhidapi-libusb.a, but as the path implies
it's for glibc. It may still work, but, again, unclear how to get it
inside the container, given that it is spun up within a single action
only.

Our best bet may be to try the hidapi-sys crate (I must have done that
in the past; it's unclear why we didn't use it) or just quick and dirty
compile the one file (?) as part of the nitrokey-sys-rs build.

As a last resort, we could use https://statifier.sourceforge.net/