Introduce dedicated minimum supported Rust version job
Introduce a dedicated CI job for checking that the program builds with
the advertised minimum supported Rust version. One problem that we have
with the existing check is that it builds the entire crate, including
all tests and other development artifacts. But we don't really care
whether anything development related adheres to our MSRV -- only the
main program and core extensions.